refactor: move BareItem enum into bare_item module
